fix: moved files to run domain
