--- src/main/kotlin/com/solve/domain/problem/dto/response/ProblemRunReponse.kt
@@ -1,5 +1,0 @@
-package com.solve.domain.problem.dto.response
-
-data class ProblemRunResponse(
-    val runId: String,
-)
