+++ src/main/kotlin/com/solve/domain/run/dto/request/RunWebSocketMessage.kt
@@ -1,0 +1,6 @@
+package com.solve.domain.run.dto.request
+
+data class RunWebSocketMessage(
+    val type: String = "",
+    val input: String? = null
+)
