+++ src/main/kotlin/com/solve/domain/run/dto/response/RunResponse.kt
@@ -1,0 +1,5 @@
+package com.solve.domain.run.dto.response
+
+data class RunResponse(
+    val runId: String,
+)
