+++ src/main/kotlin/com/solve/global/config/websocket/WebSocketConfig.kt
@@ -1,7 +1,7 @@
 package com.solve.global.config.websocket
 
-import com.solve.domain.problem.service.impl.ProblemRunService
-import com.solve.domain.problem.util.handler.CodeExecutionWebSocketHandler
+import com.solve.domain.run.service.RunService
+import com.solve.domain.run.util.handler.RunWebSocketHandler
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.socket.config.annotation.EnableWebSocket
@@ -12,10 +12,10 @@ import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry
 @EnableWebSocket
 class WebSocketConfig : WebSocketConfigurer {
     @Autowired
-    private lateinit var problemRunService: ProblemRunService
+    private lateinit var problemRunService: RunService
 
     override fun registerWebSocketHandlers(registry: WebSocketHandlerRegistry) {
-        registry.addHandler(CodeExecutionWebSocketHandler(problemRunService), "/ws/run/{runId}")
+        registry.addHandler(RunWebSocketHandler(problemRunService), "/ws/run/{runId}")
             .setAllowedOrigins("*")
     }
 }
